clipboard: refactor and add OCR support
Split the monolithic setup.go into separate files per concern
(types, clipboard, ocr). Adds tesseract-based OCR extraction
for clipboard images, gated behind the 'ocr' config flag.

Also includes several fixes uncovered during the refactor:
- initUnicodeSymbols() was never called from Setup(), so the
  IgnoreSymbols feature was completely broken
- cleanup(), saveToFile(), and trim() accessed the clipboard
  map without holding the mutex, causing data races with
  handleChange() and Query()
- saveImg() panicked on file errors instead of returning empty
- getMimetypes() used log.Println instead of slog.Error

Closes #208
